+++ one_bucket/package.json
@@ -14,6 +14,7 @@
         "@react-native-community/checkbox": "^0.5.17",
         "@react-native-masked-view/masked-view": "^0.3.1",
         "@react-navigation/bottom-tabs": "^6.5.20",
+        "@react-navigation/material-top-tabs": "^6.6.14",
         "@react-navigation/stack": "^6.3.29",
         "@reduxjs/toolkit": "^2.2.7",
         "@types/styled-components": "^5.1.34",
@@ -30,11 +31,13 @@
         "react-native-fs": "^2.20.0",
         "react-native-gesture-handler": "^2.16.1",
         "react-native-image-picker": "^7.1.2",
+        "react-native-pager-view": "^6.4.0",
         "react-native-safe-area-context": "^4.9.0",
         "react-native-screens": "^3.30.1",
         "react-native-splash-screen": "^3.3.0",
         "react-native-svg": "^12.5.1",
         "react-native-svg-transformer": "^1.5.0",
+        "react-native-tab-view": "^3.5.2",
         "react-native-toast-message": "^2.2.0",
         "react-query": "^3.39.3",
         "styled-components": "^6.1.8",
