+++ one_bucket/src/screens/home/GroupPurchase.tsx
@@ -1,0 +1,11 @@
+import { Text, View } from 'react-native'
+
+const GroupPurchase: React.FC = (): JSX.Element => {
+    return (
+        <View>
+            <Text>공동구매</Text>
+        </View>
+    )
+}
+
+export default GroupPurchase
