+++ one_bucket/src/screens/navigation/mainRoutes.tsx
@@ -20,13 +20,13 @@ import { useNavigation } from '@react-navigation/native'
 import { View } from 'react-native'
 import { TouchableOpacity } from 'react-native-gesture-handler'
 import Chat from 'screens/Chat'
-import Home from 'screens/Home'
+import HomeRoutes from 'screens/home/HomeRoutes'
 import Mypage from 'screens/Mypage'
 
 export const mainRoutes = [
     {
         name: strings.homeScreenName,
-        component: Home,
+        component: HomeRoutes,
         activeIconLight: <IcTabHomeActive fill={baseColors.SCHOOL_BG} />,
         inactiveIconLight: <IcTabHomeInactive fill={baseColors.SCHOOL_BG} />,
         activeIconDark: <IcTabHomeActive fill={baseColors.WHITE} />,
